Migrate MultibindingGraphWithWithScopedSetDeps to box test
